Extract the getUpcomingEvents function from event-series pages
